requirements.txt
beautifulsoup4>=4.12,<5
cachetools>=5.3,<6
fastapi>=0.115,<1
httpx>=0.27,<1
python-dotenv>=1.0,<2
rapidfuzz>=3.9,<4
uvicorn[standard]>=0.32,<1
ytmusicapi>=1.11,<2
